refactor(🔧/remove🔥): Delete previous game page and restructure directories
