--- src/Components/CacaoBoard/CacaoBoard.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-interface Props {}
-
-const CacaoBoard = (props: Props) => {
-  return <div className="h-8 bg-mexicanBoneLight">100 cacao</div>;
-};
-
-export default CacaoBoard;
