--- src/Containers/GameNew/GameNew.tsx
@@ -1,125 +1,0 @@
-import React, { ReactElement, ReactNode } from "react";
-import LayoutGameNew from "../../Components/LayoutGame/LayoutGameNew";
-
-interface PropsWithChildren {
-  children: ReactNode;
-}
-
-function Layout(props: PropsWithChildren): ReactElement {
-  const { children } = props;
-  return (
-    <div className="relative flex-1 flex flex-row-reverse tablet:flex-col w-full pr-20 tablet:pr-0">
-      {children}
-    </div>
-  );
-}
-function Card(props: { color: string }) {
-  const { color } = props;
-  return (
-    <div
-      className={`w-16 h-16 bg-milpaPink-dark tablet:bg-transparent tablet:w-leaderboard-card-small tablet:h-leaderboard-card-small tablet:bg-card-${color}`}
-    ></div>
-  );
-}
-
-function CardPlaceholder() {
-  return (
-    <div className="tablet:w-leaderboard-card-small tablet:h-leaderboard-card-small">
-      {" "}
-      1/16
-    </div>
-  );
-}
-
-function Cards() {
-  return (
-    <div className="fixed tablet:static right-0 top-20 flex flex-col-reverse tablet:flex-row justify-between items-center w-20 tablet:w-full p-1 tablet:p-3 bg-milpaPink-light h-mountains tablet:h-44">
-      <Card color="green" />
-      <Card color="green" />
-      <Card color="green" />
-      <Card color="green" />
-      <Card color="pink" />
-      <Card color="pink" />
-      <Card color="pink" />
-      <CardPlaceholder />
-    </div>
-  );
-}
-
-function Slot() {
-  return <div className="w-9 h-9 tablet:h-14 tablet:w-14 bg-yellow-700"></div>;
-}
-
-function EdgeTopOrBottom() {
-  return (
-    <div className="col-start-2 col-span-4 flex flex-row justify-between">
-      <Slot />
-      <Slot />
-      <Slot />
-      <Slot />
-    </div>
-  );
-}
-
-function EdgeMilpaEdge() {
-  return (
-    <>
-      <div className="col-start-1 col-end-2">
-        <Slot />
-      </div>
-      <div className="col-start-2 col-end-6 flex flex-row justify-between">
-        <Slot />
-        <Slot />
-        <Slot />
-        <Slot />
-      </div>
-      <div className="col-end-7 col-span-1 ...">
-        <Slot />
-      </div>
-    </>
-  );
-}
-
-function Milpa() {
-  return (
-    <div className="h-milpa-board-mobile w-milpa-board-mobile tablet:h-milpa tablet:w-milpa bg-yellow-800 m-3 p-1 tablet:p-2">
-      <div className="grid grid-cols-6 gap-2 tablet:gap-4 place-content-center h-full">
-        <EdgeTopOrBottom />
-        <EdgeMilpaEdge />
-        <EdgeMilpaEdge />
-        <EdgeMilpaEdge />
-        <EdgeMilpaEdge />
-        <EdgeTopOrBottom />
-      </div>
-    </div>
-  );
-}
-
-function Info() {
-  return <div className="h-80 tablet:h-milpa w-64 bg-yellow-600"></div>;
-}
-
-function Boards() {
-  return (
-    <div className="flex-auto flex flex-col tablet:flex-row justify-evenly items-center w-full tablet:p-4 bg-milpaGreen-light">
-      <Milpa />
-      <Info />
-      <Milpa />
-    </div>
-  );
-}
-
-export default function GameNew(): ReactElement {
-  const players = {
-    local: "Rolando",
-    remote: "Gabinkova",
-  };
-  return (
-    <LayoutGameNew players={players} scores={["0", "0"]} yourTurn={true}>
-      <Layout>
-        <Cards />
-        <Boards />
-      </Layout>
-    </LayoutGameNew>
-  );
-}
