--- src/Hooks/useCreateOrJoinGameContext/useCreateOrJoinGameContext.ts
@@ -1,12 +1,0 @@
-import { useContext } from "react";
-import { CreateOrJoinGameContext } from "../../Components/CreateOrJoinGameProvider/CreateOrJoinGameProvider";
-
-export default function useCreateorJoinGameContext() {
-  const context = useContext(CreateOrJoinGameContext);
-  if (!context) {
-    throw new Error(
-      "useCreateOrJoinGameContext must be used within a CreateorJoinGameProvider"
-    );
-  }
-  return context;
-}
