--- src/Hooks/useGameContext/useGameContext.ts
@@ -1,10 +1,0 @@
-import { useContext } from "react";
-import { GameContext } from "../../Components/GameProvider/GameProvider";
-
-export default function useGameContext() {
-  const context = useContext(GameContext);
-  if (!context) {
-    throw new Error("useGameContext must be used within a GameProvider");
-  }
-  return context;
-}
