+++ src/common/socket/socket.ts
@@ -1,7 +1,7 @@
 import { io } from "socket.io-client";
 import { MiSocket } from "../../Realms/Pure/types";
 
-const URL = "https://bbdc-2806-2f0-9180-a5a1-5954-e925-b5b1-8d19.ngrok.io";
+const URL = "http://localhost:3000";
 
 const newSocket = (roomCode: string, nickname: string | undefined) => {
   const socket: MiSocket = io(URL, {
