+++ src/containers/Game/Info.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Info() {
+    return <div className="h-80 tablet:h-milpa w-64 bg-yellow-600"/>;
+}
