+++ src/containers/Game/Slot.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Slot() {
+    return <div className="w-9 h-9 tablet:h-14 tablet:w-14 bg-yellow-700"/>;
+}
+
+
