+++ src/pages/WelcomePage.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+import {Welcome} from "../containers/Welcome";
+export default function WelcomePage(){
+    return (<Welcome/>)
+}
