+++ src/pages/index.ts
@@ -1,0 +1,9 @@
+import WelcomePage from './WelcomePage'
+import CreateGamePage from "./CreateGamePage";
+import GamePage from './GamePage'
+
+export {
+    WelcomePage,
+    CreateGamePage,
+    GamePage
+}
